mypage 데이터 불러와서 뿌리는거 까지 성공, 이후 가져온 데이터 가공해서 보여줘야 함.

+++ backend/src/main/java/com/noblesi/travelplanner/mapper/MyPageMapper.java
@@ -8,7 +8,7 @@ import com.noblesi.travelplanner.dto.member.MemberInfoRequest;
 @Mapper
 public interface MyPageMapper {
     
-    public MemberInfoDomain selectMemberInfo(int memberId);
+    public MemberInfoDomain selectMemberInfo(long memberId);
 
     public int updateMemberInfo(MemberInfoRequest memberInfoRequest);
